support correct formatting of multiple entries of the same tag
Using some CSS magic this hides the repeated icon for consecutive email, phone, etc. entries in the contact-list. As well as only underlining the last entry of each grouping.

Note: that this only works for consecutive entries of the same paper-contact-* tag where that tag isn't reused later in the list but it is a start. If multiple non-consecutive repeated tags are used styling bugs will occur, but ordering in this manner would be contrary to the Material UI spec and is therefore of little concern for now.